tctl resources: convert DynamicWindowsDesktop to new handler format
Updates #60370
